fixed lexer, new global rules, new specific rules word & string
