+++ includes/lexer.h
@@ -1,38 +1,44 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   minishell.h                                        :+:      :+:    :+:   */
+/*   lexer.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: rsanchez <rsanchez@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/22 16:19:25 by rsanchez          #+#    #+#             */
-/*   Updated: 2021/12/18 14:20:53 by rsanchez         ###   ########.fr       */
+/*   Updated: 2021/12/18 14:17:42 by romain           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef MINISHELL_H
-# define MINISHELL_H
+#ifndef LEXER_H
+# define LEXER_H
 
 # include "libft.h"
 # include <termios.h>
 
-typedef struct termios	t_term;
-
-typedef struct s_minishell
+enum	e_char_type
 {
-	t_gc		gc;
-	t_vector	tokens;
-	t_vector	history;
-	t_term		term_config;
-}			t_msh;
-BOOL		init_terminal(t_msh *msh);
-BOOL		reset_terminal(t_msh *msh);
-void		exit_program(t_msh *msh);
-void		*assert_gc(t_msh *msh, void *data, void (*f)(void *));
-void		*assert_malloc(t_msh *msh, void *data);
-BOOL		assert_bool(t_msh *msh, BOOL check);
-int			assert_errno(t_msh *msh, int i);
-char		*get_input(t_msh *msh);
-void		tokenizer(t_msh *msh, char *input);
+	OPERATOR = 0,
+	QUOTE,
+	CHARACTER,
+	BLANK,
+	IGNORE,
+	CHAR_NB
+};
+
+enum	e_token_type
+{
+	TOK_OPERATOR = 0,
+	TOK_QUOTED_STR,
+	TOK_WORD,
+	NOT_A_TOKEN,
+	TOK_NB
+};
+
+int			get_token(char *input);
+
+int			get_char_type(int c);
+int			get_token_type(int char_type);
+BOOL		is_allowed(int char_type, int token_type);
 
 #endif
